src/dispatcher/requirements.txt
Flask==3.1.2
requests==2.31.0
google-cloud-storage==2.15.0
google-cloud-run==0.10.0
cloudevents==1.10.1
pydantic==2.10.4
PyYAML==6.0.2
debugpy==1.8.1